Convert unitless module declaration into block form
Otherwise, it would have required a `unit` declarator and the braces would
have to have been removed.

The reason for this change is, that as of Rakudo 2015.05, the `unit`
declarator is required before using `module`, `class` or `grammar`
declarations (unless it uses a block).  Code still using the old blockless
semicolon form will throw a warning. This commit stops the warning from
appearing in the new Rakudo.